Add services for users and payment cards

+++ src/main/java/io/github/lefreshz/user_service/repository/PaymentCardRepository.java
@@ -18,7 +18,7 @@ public interface PaymentCardRepository extends JpaRepository<PaymentCard, Long>,
 
   Page<PaymentCard> findByUser_UserId(Long userId, Pageable pageable);
 
-  long countByUser_UserId(Long userId);
+  int countByUser_UserId(Long userId);
 
   boolean existsByNumber(String number);
